// src/types/moon.tsx
import type { ReactNode } from 'react';

// Formato dos dados que vêm da API
export interface AstroData {
  sunrise: string;
  sunset: string;
  moonrise: string;
  moonset: string;
  moon_phase: string;
  moon_illumination: string;
  is_moon_up: number;
  is_sun_up: number;
}

export interface AstronomyResponse {
  location: {
    name: string;
    region: string;
    country: string;
    lat: number;
    lon: number;
    tz_id: string;
    localtime_epoch: number;
    localtime: string;
  };
  astronomy: {
    astro: AstroData;
  };
}

// Formato do nosso dicionário de fases
export interface MoonPhaseInfo {
  label: string;
  icon: ReactNode;
  desc: string;
}